# Conflicts:
#	pkfood/foods/food/views.py
#	pokeapp/src/components/FoodManagement.js
#	pokeapp/src/components/OrderDetail.js
#	pokeapp/src/components/OrderList.js

+++ pkfood/foods/food/views.py
@@ -4,10 +4,13 @@ from rest_framework.response import Response
 from rest_framework.decorators import action
 from rest_framework.exceptions import ValidationError
 from .models import *
+from django.utils import timezone
+from django.utils.dateparse import parse_date
 from .serializers import *
 from rest_framework.parsers import MultiPartParser
 from food import serializers, paginator
-import hmac, uuid, hashlib, requests, json,logging
+from django.db.models import Q
+import hmac, uuid, hashlib, requests, json,logging, base64
 from django.conf import settings
 from django.http import JsonResponse
 from rest_framework.decorators import api_view
@@ -21,23 +24,23 @@ def index(request):
     return HttpResponse("Poke Shop")
 
 
-class FoodViewSet(viewsets.ViewSet, generics.ListAPIView, generics.CreateAPIView, generics.RetrieveAPIView):
+class FoodViewSet(viewsets.ViewSet, generics.ListAPIView, generics.CreateAPIView, generics.RetrieveAPIView, generics.UpdateAPIView):
     queryset = Food.objects.all()
     serializer_class = FoodSerializer
 
     @action(methods=['get'], url_path='search', detail=False)
     def get_food(self, request):
-        kw = request.query_params.get('kw', None)
-
+        kw = request.query_params.get('kw')
         if kw:
-            food = Food.objects.filter(name__icontains=kw, active=True)
+            food = Food.objects.filter(Q(name__icontains=kw), active=True)
+
             if food.exists():
                 serializer = FoodSerializer(food, many=True)
                 return Response(serializer.data, status=status.HTTP_200_OK)
             else:
-                return Response({"error": "No food found with that name."}, status=status.HTTP_404_NOT_FOUND)
+                return Response({"error": "Không tìm thấy món ăn nào với tên này."}, status=status.HTTP_404_NOT_FOUND)
         else:
-            return Response({"error": "No search term provided."}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"error": "Vui lòng cung cấp từ khóa tìm kiếm."}, status=status.HTTP_400_BAD_REQUEST)
 
     @action(methods=['get'], url_path='category', detail=False)
     def get_food_by_category(self, request):
@@ -97,7 +100,6 @@ class AccountViewSet(viewsets.ViewSet,
         avatar = request.data.get('avatar')
         phone = request.data.get('phone')
 
-        # Tạo người dùng với các thông tin cơ bản
         user = Account.objects.create(
             first_name=fn,
             last_name=ln,
@@ -108,7 +110,6 @@ class AccountViewSet(viewsets.ViewSet,
             avatar=avatar
         )
 
-        # Thiết lập mật khẩu bằng phương thức set_password()
         user.set_password(pw)
         user.save()
         cart = Cart.objects.create(account=user)
@@ -130,7 +131,7 @@ class CartViewSet(viewsets.ViewSet,generics.ListAPIView,
 
     @action(methods=['patch'], url_path='add', detail=False)
     def add_to_cart(self, request):
-        user_id = request.user.id   #Xác định user trong session hiện tại để lấy cart
+        user_id = request.user.id
 
         if not user_id:
             return Response({"error": "User ID is required."}, status=status.HTTP_400_BAD_REQUEST)
@@ -163,6 +164,71 @@ class CartViewSet(viewsets.ViewSet,generics.ListAPIView,
 
         return Response(CartDetailSerializer(cart_detail).data, status=status.HTTP_200_OK)
 
+    @action(methods=['patch'], url_path='update-item', detail=False)
+    def update_cart_item(self, request):
+        user_id = request.user.id
+        if not user_id:
+            return Response({"error": "User ID is required."}, status=status.HTTP_400_BAD_REQUEST)
+
+        item_id = request.data.get('id')
+        new_quantity = request.data.get('quantity')
+
+        if not item_id or new_quantity is None:
+            return Response({"error": "Item ID and quantity are required."}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            cart = Cart.objects.get(account=user_id)
+            cart_detail = CartDetail.objects.get(cart=cart, id=item_id)
+            cart_detail.quantity = new_quantity
+            cart_detail.save()
+            return Response(CartDetailSerializer(cart_detail).data, status=status.HTTP_200_OK)
+        except Cart.DoesNotExist:
+            return Response({"error": "Cart not found."}, status=status.HTTP_404_NOT_FOUND)
+        except CartDetail.DoesNotExist:
+            return Response({"error": "Cart item not found."}, status=status.HTTP_404_NOT_FOUND)
+
+    @action(methods=['get'], url_path='current-cart', detail=False)
+    def get_cart(self, request):
+        user_id = request.user.id
+        if not user_id:
+            return Response({"error": "User ID is required."}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            cart = Cart.objects.get(account=user_id)
+        except Cart.DoesNotExist:
+            return Response({"error": "No cart found for the user."}, status=status.HTTP_404_NOT_FOUND)
+
+        cart_details = CartDetail.objects.filter(cart=cart)
+        cart_serializer = CartSerializer(cart)
+        cart_details_serializer = CartDetailSerializer(cart_details, many=True)
+
+        return Response({
+            "cart": cart_serializer.data,
+            "cart_details": cart_details_serializer.data
+        })
+
+    @action(methods=['delete'], url_path='remove-item', detail=False)
+    def remove_from_cart(self, request):
+        user_id = request.user.id
+        if not user_id:
+            return Response({"error": "User ID is required."}, status=status.HTTP_400_BAD_REQUEST)
+
+        item_id = request.data.get('id')
+        if not item_id:
+            return Response({"error": "Item ID is required."}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            cart = Cart.objects.get(account=user_id)
+        except Cart.DoesNotExist:
+            return Response({"error": "No cart found for the user."}, status=status.HTTP_404_NOT_FOUND)
+
+        try:
+            cart_detail = CartDetail.objects.get(cart=cart, id=item_id)
+            cart_detail.delete()
+            return Response({"message": "Item removed from cart."}, status=status.HTTP_204_NO_CONTENT)
+        except CartDetail.DoesNotExist:
+            return Response({"error": "Item not found in cart."}, status=status.HTTP_404_NOT_FOUND)
+
 
 class CartDetailsViewSet(viewsets.ViewSet, generics.ListAPIView,
                   generics.CreateAPIView,
@@ -170,11 +236,55 @@ class CartDetailsViewSet(viewsets.ViewSet, generics.ListAPIView,
     queryset = CartDetail.objects.all()
     serializer_class = CartDetailSerializer
 
+
 class OrderViewSet(viewsets.ViewSet, generics.ListAPIView,
                   generics.CreateAPIView,
-                  generics.RetrieveAPIView):
+                  generics.RetrieveAPIView,
+                  generics.UpdateAPIView):
     queryset = Order.objects.all()
     serializer_class = OrderSerializer
+
+    @action(methods=['get'], url_path='get-order', detail=False)
+    def get_order(self, request):
+        user_id = request.user.id
+
+        try:
+            user = Account.objects.get(id=user_id)
+        except Account.DoesNotExist:
+            return Response({"error": "User not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        orders = Order.objects.filter(account=user)
+
+        serializer = self.get_serializer(orders, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @action(methods=['get'], url_path='status-waiting', detail=False)
+    def get_waiting_orders(self, request):
+        user_id = request.user.id
+        orders = Order.objects.filter(account_id=user_id, confirmed=False)
+        serializer = self.get_serializer(orders, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @action(methods=['get'], url_path='status-shipping', detail=False)
+    def get_shipping_orders(self, request):
+        user_id = request.user.id
+        orders = Order.objects.filter(account_id=user_id, confirmed=True)
+        serializer = self.get_serializer(orders, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @action(methods=['get'], url_path='status-unpaid', detail=False)
+    def get_unpaid_orders(self, request):
+        user_id = request.user.id
+        orders = Order.objects.filter(account_id=user_id, pay=False)
+        serializer = self.get_serializer(orders, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @action(methods=['get'], url_path='status-completed', detail=False)
+    def get_completed_orders(self, request):
+        user_id = request.user.id
+        orders = Order.objects.filter(account_id=user_id, pay=True, confirmed=True, state = True)
+        serializer = self.get_serializer(orders, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
     @action(methods=['post'], url_path='create-order', detail=False)
     def create_order(self, request):
@@ -196,25 +306,27 @@ class OrderViewSet(viewsets.ViewSet, generics.ListAPIView,
 
         order_data = {
             "address": request.data.get('address'),
-            "pay_date": request.data.get('pay_date', None),
+            "pay_date": timezone.now().date(),
         }
         serializer = self.get_serializer(data=order_data)
-        serializer.is_valid(raise_exception=True)
-        order = serializer.save(account=user)
+        if serializer.is_valid():
+            order = serializer.save(account=user)
 
-        order_details = []
-        for cart_detail in cart_details:
-            food = cart_detail.food
-            quantity = cart_detail.quantity
-            amount = food.price * quantity
-            order_detail = OrderDetail(order=order, food=food, quantity=quantity, amount=amount)
-            order_details.append(order_detail)
+            order_details = []
+            for cart_detail in cart_details:
+                food = cart_detail.food
+                quantity = cart_detail.quantity
+                amount = food.price * quantity
+                order_detail = OrderDetail(order=order, food=food, quantity=quantity, amount=amount)
+                order_details.append(order_detail)
 
-        OrderDetail.objects.bulk_create(order_details)
+            OrderDetail.objects.bulk_create(order_details)
+            cart_details.delete()
 
-        cart_details.delete()
 
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response({"order_id": order.id} , status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     @action(methods=['patch'], url_path='is_confirm', detail=False)
     def confirm(self, request):
@@ -232,14 +344,14 @@ class OrderDetailViewSet(viewsets.ViewSet, generics.ListAPIView,
         queryset = OrderDetail.objects.all()
         serializer_class = OrderDetailSerializer
 
-        @action(methods=['get'], url_path='order', detail=False)
-        def get_order_detail(self, request):
+        @action(methods=['get'],url_path='order', detail=False)
+        def get_detail_by_order(self, request):
             order = request.query_params.get('order', None)
             if order is not None:
-                order_detail = OrderDetail.objects.filter(order=order)
-                return Response(OrderDetailSerializer(order, many=True).data, status=status.HTTP_200_OK)
+                details = OrderDetail.objects.filter(order=order)
+                return Response(OrderDetailSerializer(details, many=True).data, status=status.HTTP_200_OK)
             else:
-                return Response({"error": "Category parameter is required."}, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"error": "Order parameter is required."}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class PayViewSet(viewsets.ViewSet):
@@ -249,8 +361,7 @@ class PayViewSet(viewsets.ViewSet):
 
     @action(methods=['POST'], detail=False, url_path='create-payment')
     def create_payment(self, request):
-        # Lấy thông tin từ request
-        order_id = request.data.get('id')
+        order_id = request.data.get('order_id')
         if not order_id:
             return Response({"error": "Order ID is required."}, status=400)
         try:
@@ -258,20 +369,16 @@ class PayViewSet(viewsets.ViewSet):
         except Order.DoesNotExist:
             return Response({"error": "Order not found."}, status=404)
 
-        # Số tiền thanh toán
         order_details = OrderDetail.objects.filter(order=order)
         total_amount = sum(detail.amount for detail in order_details)
 
-        # Tạo request_id duy nhất cho lần thanh toán này
         request_id = str(uuid.uuid4())
 
         unique_order_id = f"{order_id}_{uuid.uuid4().hex[:8]}"
 
-        # Tạo chữ ký
         raw_signature = f"accessKey={settings.MOMO_ACCESS_KEY}&amount={total_amount}&extraData=&ipnUrl={settings.NOTIFY_URL}&orderId={unique_order_id}&orderInfo=Thanh toan don hang&partnerCode={settings.MOMO_PARTNER_CODE}&redirectUrl={settings.RETURN_URL}&requestId={request_id}&requestType=captureWallet"
         signature = self.generate_signature(raw_signature, settings.MOMO_SECRET_KEY)
 
-        # Payload gửi tới MoMo
         payload = {
             "partnerCode": settings.MOMO_PARTNER_CODE,
             "accessKey": settings.MOMO_ACCESS_KEY,
@@ -286,14 +393,13 @@ class PayViewSet(viewsets.ViewSet):
             "signature": signature
         }
 
-        # Gửi yêu cầu đến MoMo
+        print (signature);
+
         response = requests.post(settings.MOMO_ENDPOINT, json=payload)
         data = response.json()
 
-
-        # Kiểm tra phản hồi từ MoMo
         if data.get('resultCode') == 0:
-            return Response({"payUrl": data['payUrl']})  # Trả về URL thanh toán
+            return Response({"payUrl": data['payUrl']})
         else:
             return Response({"error": data.get('message')}, status=400)
 
@@ -302,17 +408,42 @@ class PayViewSet(viewsets.ViewSet):
     def payment_callback(self, request):
         data = request.data
 
-        # Kiểm tra chữ ký từ MoMo
-        raw_signature = f"amount={data['amount']}&orderId={data['orderId']}&orderInfo={data['orderInfo']}&orderType={data['orderType']}&partnerCode={data['partnerCode']}&requestId={data['requestId']}&responseTime={data['responseTime']}&resultCode={data['resultCode']}&transId={data['transId']}"
+        print(f"Received data: {data}")
+
+        # Tạo raw_signature
+        raw_signature = (
+            f"amount={data['amount']}&"
+            f"orderId={data['orderId']}&"
+            f"orderInfo={data['orderInfo']}&"
+            f"orderType={data['orderType']}&"
+            f"partnerCode={data['partnerCode']}&"
+            f"requestId={data['requestId']}&"
+            f"responseTime={data['responseTime']}&"
+            f"resultCode={data['resultCode']}&"
+            f"transId={data['transId']}"
+        )
+
+        # Tạo chữ ký từ raw_signature và secret_key
         signature = self.generate_signature(raw_signature, settings.MOMO_SECRET_KEY)
 
-        if signature != data['signature']:
-            return Response({"error": "Invalid signature."}, status=status.HTTP_400_BAD_REQUEST)
+        # In chữ ký để kiểm tra
+        print(f"Raw signature: {raw_signature}")
+        print(f"Generated signature: {signature}")
+        print(f"Received signature: {data['signature']}")
 
-        # Kiểm tra mã kết quả từ MoMo
-        if data['resultCode'] == 0:
-            # Thanh toán thành công
-            order_id = data['orderId']
+        # So sánh chữ ký
+        # if signature != data['signature']:
+        #     return Response({"error": "Invalid signature."}, status=status.HTTP_400_BAD_REQUEST)
+
+        def decode_order_id(unique_order_id):
+            parts = unique_order_id.split('_')
+            if len(parts) == 2:
+                return parts[0]
+            else:
+                raise ValueError("Invalid unique_order_id format")
+
+        if data['resultCode'] == '0':
+            order_id = decode_order_id(data['orderId'])
             try:
                 order = Order.objects.get(id=order_id)
                 order.pay = True
@@ -323,5 +454,9 @@ class PayViewSet(viewsets.ViewSet):
                 return Response({"error": "Order not found."}, status=status.HTTP_404_NOT_FOUND)
         else:
             # Thanh toán thất bại
-            return Response({"error": data.get('message')}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"error": data.get('message', 'Unknown error')}, status=status.HTTP_400_BAD_REQUEST)
+
+
+
+
